shell(install): Add installation script that symlinks the files
It removes the extension, adds the `,` prefix and symlinks to
~/.local/bin.

It would be better if I had a separate directory for these so I could
nuke it before install. That way, I uninstall/rename old ones. For now,
though, this will do.

I say that last sentence a lot, and then immediately go back on it.